# Vectorize implementation of compute concept loss for CBM model. Free speed up!