This change replaces the concept loss logic computation from the CBM with a vectorized implementation. It is a no diff in terms of metrics as you can see below. But it does make this loss computation significantly faster

I wrote some profiling code in another file I won't contribute:

Using the same numbers as CBM, batch_size=64, num_concepts=112

compute_concept_loss_og time over 1000 runs: 5.444659s
compute_concept_loss_vectorized time over 1000 runs: 0.057705s

So it is a lot faster

Even on the end to end 300 epochs of training it is several minutes (>4) of savings on my machine



```
---------------- vectorized 300 no double

TRAINING FINISHED

EVALUATION ON THE TEST SET:

Epoch 300: 100%|████████████████████████████████████████████████████████████████████████| 91/91 [00:23<00:00,  3.92it/s]
Test: target_loss: 1.567 prec_loss: 0.000 concepts_loss: 16.392 total_loss: 17.959 y_accuracy: 0.699 c_accuracy: 0.951 complete_c_accuracy: 0.449 target_jaccard: 0.537 concept_jaccard: 0.776 y_AUROC: 0.986 y_AUPR: 0.748 y_Brier: 0.438 y_ECE: 0.123 y_TL-ECE: 0.138 c_AUROC: 0.977 c_AUPR: 0.921 c_Brier: 0.038 c_ECE: 0.024 c_TL-ECE: 0.042


real    82m29.957s

---------------- original 300
TRAINING FINISHED

EVALUATION ON THE TEST SET:

Epoch 300: 100%|████████████████████████████████████████████████████████████████████████| 91/91 [00:35<00:00,  2.53it/s]
Test: target_loss: 1.567 prec_loss: 0.000 concepts_loss: 16.392 total_loss: 17.959 y_accuracy: 0.699 c_accuracy: 0.951 complete_c_accuracy: 0.449 target_jaccard: 0.537 concept_jaccard: 0.776 y_AUROC: 0.986 y_AUPR: 0.748 y_Brier: 0.438 y_ECE: 0.123 y_TL-ECE: 0.138 c_AUROC: 0.977 c_AUPR: 0.921 c_Brier: 0.038 c_ECE: 0.024 c_TL-ECE: 0.042


real    86m47.389s
```